1 - Consume Grpc service from Basket as a client 2 - Modify Basket to use grpc by injecting proto service 3 - Configure in IOC the Grpc

+++ src/Services/Basket/Basket.API/Models/ShoppingCart.cs
@@ -7,7 +7,7 @@
 
         public List<ShoppingCartItem> Items { get; set; } = new ();
 
-        public decimal TotalPrice => Items.Sum(x => x.Price * x.Quantity);
+        public double TotalPrice => Items.Sum(x => x.Price * x.Quantity);
 
 
         public ShoppingCart(string userName)
