+++ src/Services/Basket/Basket.API/Models/ShoppingCartItem.cs
@@ -8,7 +8,7 @@
 
         public string Color { get; set; } = default!;
 
-        public decimal Price { get; set; } = default!;
+        public double Price { get; set; } = default!;
 
         public string ProductName { get; set; } = default!;
